test: refactor logic to add/remove extended resources
